Renamed DynamicActionInfoService to DynamicActionInfoStore
